201.Scripts/requirements.txt
nltk==3.8.1
sklearn==0.0
pandas==2.0.2
gst_calculation==0.1.2
transformers==4.28.1
pytorch-nlp==0.5.0
pytorch-pretrained-bert